move division plugins to arithmetic.ts
